fixes #22 items can now be built in their own files and loaded dynamically
