+++ index.html
@@ -2,7 +2,7 @@
 <html lang="en">
 <head>
   <meta charset="UTF-8">
-  <title>Demo</title>
+  <title>Wake</title>
   <link rel="stylesheet" href="dist/base.css">
   <link rel="stylesheet" href="./node_modules/perfect-scrollbar/dist/css/perfect-scrollbar.min.css">
   <style type="riot"></style>
